All Major Update

+++ src/main/java/com/vlazma/Controllers/ProductController.java
@@ -1,6 +1,8 @@
 package com.vlazma.Controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,6 +15,9 @@ import org.springframework.web.bind.annotation.RestController;
 import com.vlazma.Dto.Product.ProductRequest;
 import com.vlazma.Services.ProductService;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 
 @RestController
@@ -21,52 +26,65 @@ public class ProductController {
     @Autowired
     private ProductService productService;
 
+    @Operation(summary = "Showing Product Data", security = { @SecurityRequirement(name = "bearer-key") })
     @GetMapping("/")
-    public Object get(){
+    public Object get(HttpServletRequest request) {
+        if (request.isUserInRole("ROLE_CUSTOMER")) {
+            return productService.getAllAvailableProduct();
+        }
         return productService.getAllProducts();
     }
 
-    @GetMapping("/available-product")
-    public Object getAvailable(){
-        return productService.getAllAvailableProduct();
-    }
-
     @GetMapping("/{id}")
-    public Object findById(@PathVariable int id){
+    public Object findById(@PathVariable int id) {
         return productService.findById(id);
     }
 
+    @Operation(summary = "Showing ProductData by category id", security = { @SecurityRequirement(name = "bearer-key") })
     @GetMapping("/category-id/{id}")
-    public Object getByCategoryId(@PathVariable int id){
+    public Object getByCategoryId(HttpServletRequest httpServletRequest, @PathVariable int id) {
         return productService.getAllProductsByCategoryId(id);
     }
 
+    @Operation(summary = "Showing ProductData by name", security = { @SecurityRequirement(name = "bearer-key") })
     @GetMapping("/category-name/{name}")
-    public Object getByCategoryName(@PathVariable String name){
+    public Object getByCategoryName(HttpServletRequest request, @PathVariable String name) {
         return productService.getAllProductsByCategoryName(name);
     }
 
+    @Operation(summary = "Showing ProductData by Category Name", security = {
+            @SecurityRequirement(name = "bearer-key") })
     @GetMapping("/name-contains/{like}")
-    public Object getByNameContains(@PathVariable String like){
+    public Object getByNameContains(HttpServletRequest request, @PathVariable String like) {
         return productService.getAllProductsContains(like);
     }
 
+    @Operation(summary = "Adding Product Data", security = { @SecurityRequirement(name = "bearer-key") })
     @PostMapping("/")
-    public Object create(@Valid@RequestBody ProductRequest productRequest,Errors errors){
-        return productService.create(productRequest, errors);
+    public Object create(HttpServletRequest request, @Valid @RequestBody ProductRequest productRequest, Errors errors) {
+        if (request.isUserInRole("ROLE_ADMIN")) {
+            return productService.create(productRequest, errors);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Only Admin Can Acces It");
     }
 
-    @PostMapping("/{id}")
-    public Object edit(@PathVariable int id,@Valid@RequestBody ProductRequest productRequest,Errors errors){
-        return productService.edit(id, productRequest, errors);
+    @Operation(summary = "Showing ProductData by name", security = { @SecurityRequirement(name = "bearer-key") })
+    @PostMapping("/edit/{id}")
+    public Object edit(HttpServletRequest request, @PathVariable int id,
+            @Valid @RequestBody ProductRequest productRequest, Errors errors) {
+        if (request.isUserInRole("ROLE_ADMIN")) {
+            return productService.edit(id, productRequest, errors);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Only Admin Can Acces It");
     }
 
-    @DeleteMapping("/{id}")
-    public Object delete(@PathVariable int id){
-        return productService.delete(id);
+    @Operation(summary = "Showing ProductData by name", security = { @SecurityRequirement(name = "bearer-key") })
+    @DeleteMapping("/delete/{id}")
+    public Object delete(HttpServletRequest request, @PathVariable int id) {
+        if (request.isUserInRole("ROLE_ADMIN")) {
+            return productService.delete(id);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Only Admin Can Acces It");
     }
-
-
-
 
 }
